### 2020-08-14 #2568 @whedon
[PRE REVIEW]: dorado: A Python package for simulating passive particle transport in shallow-water flows
**Submitting author:** @elbeejay (<a href="http://orcid.org/0000-0002-1343-193X">Jayaram Hariharan</a>)
**Repository:** <a href="https://github.com/passaH2O/dorado" target ="_blank">https://github.com/passaH2O/dorado</a>
**Version:** v1.0.0
