### 2020-08-19 #2568 @kbarnhart
@elbeejay I've now completed my pre-review checks and will begin looking for reviewers. I found one minor item that I think would be good to address before review begins: you have tests, but no instructions for how to run them. I would recommend indicating how a potential users would test their install somewhere on the readme and/or docs.
