### 2020-08-19 b9e8a2d
info about running unit tests
