### 2020-08-19 #2568 @kbarnhart
:wave: @dbuscombe-usgs @gassmoeller @TomasBeuzen would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

This submission is for "dorado: A Python package for simulating passive particle transport in shallow-water flows". I'm reaching out to you based on your expertise in fluid flow, particle methods, and/or python.

This is a pre-review issue which is used to find reviewers. Once 2-3 reviewers have been found I'll start the review on a dedicated GitHub issue. At present we are asking reviewers to complete reviews in 6 weeks. If you are not able to review and have someone to recommend, please mention them here (when mentioning, please place a space after the @ of a github handle, for example, you would refer to me as "@ kbarnhart").

If you are interested in reviewing, I would recommend looking over the journal's [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) before accepting the review invitation.

If you have any questions about the JOSS review process, please do not hesitate to reach out to me by commenting on this issue or emailing me directly at krbarnhart@usgs.gov
