### 2020-08-20 #2568 @TomasBeuzen
@gassmoeller - sure no preference, I'm happy with whatever @kbarnhart decides. FYI I'm more experienced with Python than hydrodynamics so my review will focus mostly on the code (style, usability, functionality, etc.) and clarity of the manuscript.
