### 2020-08-20 #2585 @whedon
[REVIEW]: dorado: A Python package for simulating passive particle transport in shallow-water flows
**Submitting author:** @elbeejay (<a href="http://orcid.org/0000-0002-1343-193X">Jayaram Hariharan</a>)
**Repository:** <a href="https://github.com/passaH2O/dorado" target ="_blank">https://github.com/passaH2O/dorado</a>
**Version:** v2.1.0
