### 2020-09-14 #2585 @dbuscombe-usgs
when I'm done with my review, do I just paste it here? I see the 'checklist' but the instructions are vague on the actual review submission process
