### 2020-09-14 #2585 @dbuscombe-usgs
I confirm that I have read the JOSS conflict of interest policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.

I confirm that I read and will adhere to the JOSS code of conduct.

Paper:
•	Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided? Yes
•	A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is? Yes
•	State of the field: Do the authors describe how this software compares to other commonly-used packages? Yes
•	Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)? Yes
•	References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax? Yes

The statement of need should lead with a definition of the term ‘particle tracking’. It is clear to me that this is shorthand for ‘tracking of water – including particles – in a Lagrangian framework’, but it could be clearer to the reader.
I also recommend adding the theory section in the documentation to the paper (https://passah2o.github.io/dorado/background/index.html). It is well written and answers many outstanding questions I have when left with when reading the paper on its own.

Documentation:
•	A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is? yes
•	Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution. Yes
•	Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems). Yes
•	Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)? Yes
•	Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified? Yes
•	Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

No. More instruction for how to make a pull request (e.g. https://docs.fast.ai/dev-setup), and how to submit an issue (including what information should be listed when doing so) is required. Also, provide a direct link to the github repo here

What is the motivation for maintaining both pip and conda versions? Seems like picking one and ensuring support would be easier. Also, I suggest you recommend users make use of a virtual environment (venv) or conda environment to make installation, maintenance and compatibility with other projects easier.
The demo on https://passah2o.github.io/dorado/quickstart/index.html#demo-1-using-the-high-level-api is missing:
“import matplotlib.pyplot as plt”
Also, please provide some higher level explanation for what this example is actually doing and what this is a simulation of.

General checks
•	Repository: Is the source code for this software available at the repository url? Yes
•	License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license? Yes
•	Contribution and authorship: Has the submitting author made major contributions to the software? Does the full list of paper authors seem appropriate and complete? Yes

Functionality
•	Installation: Does installation proceed as outlined in the documentation? Yes
•	Functionality: Have the functional claims of the software been confirmed? Yes

Are there any plans to add or change functionality in the future? If so, a roadmap would be a good addition. It would also be a good way for others could see themselves contributing

Instead of using the future package, presumably to make things backwards-compatible, I would suggest catering ONLY to python 3 and later. This is a new program so backwards-compatibility is certainly not expected and probably not necessary.

Great coding and commenting!

Accept with these minor comments
