+++ docs/source/examples/index.rst
@@ -22,3 +22,14 @@ Examples of the high-level API functionality are provided here. These examples r
    example11
    example12
    example13
+
+
+External Examples
+-----------------
+
+Below are links to examples of `dorado` being used in other projects.
+If you have used `dorado` as part of your project and have an example or graphic hosted online, please let us know and we will add a link to it here.
+
+`Tides on a simple 2D field <https://espin-2020.github.io/CoastalTeam/examples/HomogeneousField2D/HomogeneousField2D.html>`_ - an example of coupling `dorado` with `Landlab <https://landlab.github.io>`_.
+
+`Tides on a randomly vegetated field <https://espin-2020.github.io/CoastalTeam/examples/RandomField2D/RandomField2D.html>`_ - another example of coupling `dorado` with `Landlab <https://landlab.github.io>`_.
