### 2020-09-15 d7151f9
remove conda install
