### 2020-09-15 #2585 @kbarnhart
@dbuscombe-usgs thanks for the review (and sorry to miss this yesterday). You should be able to check the checkboxes... I'll re-invite you just in case you don't have check-box privileges, and if that doesn't work, let me know.

Sorry that exactly how to provide the review was unclear. Pasting your comments in as you have done is just fine. For more substantial comments we recommend that reviewers open an issue in the main repository. I've now copy-pasted your comments into in-repo issues. This just helps keep track of what changes in-repo are associated with the review.

It looks like @elbeejay has already started addressing these issues. 👍
