### 2020-09-15 passaH2O/dorado#7 @kbarnhart
JOSS review: Improve statement of need
From @dbuscombe-usgs's JOSS review [link to comment](https://github.com/openjournals/joss-reviews/issues/2585#issuecomment-692365791)

openjournals/joss-reviews/issues/2585

> The statement of need should lead with a definition of the term ‘particle tracking’. It is clear to me that this is shorthand for ‘tracking of water – including particles – in a Lagrangian framework’, but it could be clearer to the reader.
I also recommend adding the theory section in the documentation to the paper (https://passah2o.github.io/dorado/background/index.html). It is well written and answers many outstanding questions I have when left with when reading the paper on its own.
