### 2020-09-15 passaH2O/dorado#8 @kbarnhart
JOSS Review: Improve community guidelines
From @dbuscombe-usgs's JOSS review [link to comment](https://github.com/openjournals/joss-reviews/issues/2585#issuecomment-692365791)

openjournals/joss-reviews/issues/2585

> More instruction for how to make a pull request (e.g. https://docs.fast.ai/dev-setup), and how to submit an issue (including what information should be listed when doing so) is required. Also, provide a direct link to the github repo here
