### 2020-09-15 #2585 @kbarnhart
@dbuscombe-usgs I've gone ahead and checked all the boxes except for community guidelines (where I think you indicated "no"). You should be able to check them with your mouse. Its not critical though, as you indicated your responses in your comment.

Other than affirming that @elbeejay has addressed your comments at the very end of the review processes you are done. Thanks for a speedy and thoughtful review.
