### 2020-09-15 #2585 @dbuscombe-usgs
I'm not able to check the boxes, on my Brave browser on linux
