### 2020-09-15 #2585 @kbarnhart
@arfon FYI. If we expect github md checklists to be checkable on Linux + Brave we have a report that they are not check-able. Obviously a minor point, but a datapoint nonetheless.
