### 2020-09-15 #2585 @elbeejay
Yes we have gotten a bit of a start on addressing those comments. Thanks for the review @dbuscombe-usgs - will let you know once we've addressed everything.
