### 2020-09-25 passaH2O/dorado#11 @gassmoeller
JOSS Review - Warnings with parallel example
Related to openjournals/joss-reviews#2585

I get the attached warnings when running `examples/parallel_routing.py`. May just be a minor version issue on my machine, but I just wanted to know what you think about them.

My package versions:
python 3.8.2 (default install on ubuntu 20.04)
numpy 1.17.4
matplotlib 3.1.2
scipy 1.3.3
future 0.18.2
tqdm 4.49.0

[log.txt](https://github.com/passaH2O/dorado/files/5284043/log.txt)
