### 2020-09-25 #2585 @gassmoeller
I have reviewed the software and opened a few issues for minor points as linked above. I have some concerns about the code structure that I raise below, but the submission itself was well prepared and the code is easy to read and seems well tested. The software is useful and the paper is a nice summary and fits into JOSS's scope. I would like to have a conversation about the points I raise below, and which ones you can easily address and which have to be left for future work, but after addressing the main points I am open to recommending this submission for publication.

- You provide a clear statement of need, but I would prefer if you could specify the target audience a bit more clearly. It seems you are aiming at small- to medium-scale particle models and rapid development and use rather than high-performance and massively parallel models, is that correct? If so, please state.

- Your main class that handles all particles is called "Particle". When I looked into the code I expected this to be a class for one single particle. In terms of nomenclature, maybe something like ParticleHandler, or ParticleManager seems more appropriate (I do not insist on those names, I only suggested them, because in my own particle implementation I chose the name ParticleHandler). At least I would consider using "Particles" to make clear that it is a class that holds more than one particle. I realize this may be a disruptive change, and I do not know how much backward compatibility you have to consider, but I would like to hear your opinion on the topic.

- Your current code works for your purpose and the examples you show, but I am slightly concerned about the future maintainability of the code, because you have mixed responsibilities and an unclear class hierarchy for some of your functions. I will give some examples for your main functions and why that could become a problem later below. I would not consider this a roadblock for the publication in JOSS, but I would like to see as much as reasonable of this fixed at some point and I would like to hear your opinion on the matter:

  - You define particle locations either based on an initial seed location, or based on existing particle locations either from a previous run, or created by the user. But the initial seed location is expected to be handed over during construction of Particle, while existing locations are handed over to run_iteration. This is a bit confusing for a newcomer, but more importantly: It shows that you are not sure who (which function) is responsible for the particle generation. I would suggest to have one function specific for particle generation, e.g. `generate_particles(x_seed, y_seed, start_xindices, start_yindices)`. In a previous particle system I have written we separated that into a set of different functions (Particles::Generators::regular_reference_locations, Particles::Generators::probabilistic_locations, ..., see [here](https://github.com/dealii/dealii/blob/master/include/deal.II/particles/generators.h) if you are interested in details), but that would be overkill for your application here. Still you want to keep that option instead of spreading the responsibility for generating over multiple functions or making your `run_iteration` function into a `somehow_generate_particles_and_run_iteration` function. This would become a problem once you start to add more particle generation options (e.g. probabilistic particle distribution with varying density functions, particle patterns like discs/circles/lines).

  - A related comment about assigning responsibilities: I do not understand why the `run_iteration` function is a member of `Particle`, but the `single_iteration` function is a member of `Tools`. `single_iteration` does the core work of moving the particles, and so I would suspect it to be a member of `Particle`. It is true that `single_iteration` uses many of the `Tools` functions, but that does not mean it is conceptually a tool itself, and I would move it to `Particle`. On the other hand the functions `coord2ind`, `ind2coord`, and `unstruc2grid` are typical tools functions and I would expect them in the `Tools` class, and not as stand-alone functions in `particle_track.py`. More about `Tools` below.

  - Currently `Tools` defines a number of utility functions for the `Particle` class, and then `Particle` is derived from `Tools`. But this is not a typical `is-a` parent-child relation. `Particle` is *not* a `Tools`. `Particle` uses `Tools`, so instead it should import `Tools`, or be given an object of type `Tools` upon creation or similar. This is already a problem in your code, because functions in `Tools` use member variables that are only read in and initialized in `Particle` (e.g. self.velocity in Tools.calc_travel_times). In other words `Tools` can not be used without `Particle` or used as a base class for anything else. Utility functions or objects should usually not own any member variables (and in particular not rely on member variables of derived classes), and instead be handed all necessary parameters to compute the little computation they are supposed to do.

  - Related to the last comment: Your documentation in `Tools` currently states that `Particle` is derived from `Tools` and uses its functions, but if that would change in the future you would likely only update the `Particle` class and forget to change the documentation of `Tools` (not implying you are not careful, this is just what experience in programming shows). Instead this documentation should be in the `Particle` class, explaining why `Particle` is derived from `Tools`, this way if the relationship changes you also update the documentation. Documentation should always state what this thing does, not what others do with it.
