### 2020-09-25 #2585 @kbarnhart
@gassmoeller thanks so much for your thoughtful review. 👍

A quick note to @elbeejay and co-authors. As you address these issues I'd recommend you start an in-repo issue, and corresponding with @gassmoeller about the details there. That way the multiple comments can be addressed on separate threads.
