### 2020-09-28 passaH2O/dorado#7 @elbeejay
@gassmoeller left a comment on the statement as well ([link to original comment](https://github.com/openjournals/joss-reviews/issues/2585#issuecomment-699174900)):

> You provide a clear statement of need, but I would prefer if you could specify the target audience a bit more clearly. It seems you are aiming at small- to medium-scale particle models and rapid development and use rather than high-performance and massively parallel models, is that correct? If so, please state.
