### 2020-09-28 passaH2O/dorado#12 @elbeejay
JOSS Review: "Particle" class name
This was a comment made by @gassmoeller [here](https://github.com/openjournals/joss-reviews/issues/2585#issuecomment-699174900):

> Your main class that handles all particles is called "Particle". When I looked into the code I expected this to be a class for one single particle. In terms of nomenclature, maybe something like ParticleHandler, or ParticleManager seems more appropriate (I do not insist on those names, I only suggested them, because in my own particle implementation I chose the name ParticleHandler). At least I would consider using "Particles" to make clear that it is a class that holds more than one particle. I realize this may be a disruptive change, and I do not know how much backward compatibility you have to consider, but I would like to hear your opinion on the topic.
