### 2020-09-28 passaH2O/dorado#12 @elbeejay
We appreciate the insightful review and this comment prompted a productive discussion about the overall naming scheme for several of the classes and functions in the `dorado` code base.

Our plan is to change `Particle` to `Particles` to acknowledge that this class holds a *set* of particles not just an individual one. In addition, we plan to rename `single_iteration` to `particle_stepper` to be more descriptive and indicate that the purpose of the function is to move the particles around. We are intend to rename `params` to `modelParams` to indicate that the parameters defined in that class are related to the "global" model and not a particular single particle or set of particles.

Similarly, we plan to rename `particle_tools.py` to `LagrangianWalker.py` to indicate that the purpose of the file is to hold the functions related to the Lagrangian transport of the particles.

@gassmoeller - do you think these changes would improve the clarify of the classes and functions? Thank you again for the very helpful suggestion!
