### 2020-09-28 4850dc1
make data paths relative to the script, not terminal working directory
