### 2020-09-28 e249f6d
add os import to example script
