### 2020-09-28 e4c5737
add note about which examples must be run from the "examples" subdirectory, and make some file paths OS-agnostic in docs
