+++ docs/source/examples/example07.rst
@@ -13,16 +13,17 @@ First run :ref:`example02`, and then we can access and use the particle travel i
 
    >>> import numpy as np
    >>> import json
+   >>> import os
    >>> from dorado.routines import draw_travel_path
    >>> data = np.load('ex_deltarcm_data.npz')
    >>> depth = data['depth']
-   >>> all_walk_data = json.load(open('steady_deltarcm_example/data/data.txt'))
+   >>> all_walk_data = json.load(open('steady_deltarcm_example'+os.sep+'data'+os.sep+'data.txt'))
 
 Now we will apply the `draw_travel_path` function. In this example we are going to visualize the travel paths of the first 4 particles, indicated by the list we are providing to the `draw_travel_path` function.
 
 .. doctest::
 
    >>> draw_travel_path(depth, all_walk_data, [0, 1, 2, 3],
-   >>>                  'steady_deltarcm_example/data/travel_paths.png')
+   >>>                  'steady_deltarcm_example'+os.sep+'data'+os.sep+'travel_paths.png')
 
 .. image:: images/example07/travel_paths.png
