### 2020-09-28 passaH2O/dorado#10 @elbeejay
Hi @gassmoeller, thanks for bringing this up. We had (wrongly) assumed that examples would be called from *within* the example directory, so something like `cd examples` then `python steady_anuga_particles.py` as opposed to calling `python examples/steady_anuga_particles.py` from the project root.

But this is a valid point, and it does make sense that the user should be able to run the examples from any directory they wish, e.g. from the root: `python examples/steady_anuga_particles.py`. This has been resolved in #14 by setting paths to the example data files using the location of the example script (as opposed to the current working directory). Additional notes and information in the example documentation is being added as well to clearly state if there are limitations on where/how a given example can be run.
