### 2020-10-02 76183c7
add automated testing of example py scripts (except animate)
