+++ .github/workflows/build.yml
@@ -69,6 +69,46 @@ jobs:
       run: |
         pytest --cov=dorado/ --cov-report=xml
 
+  examples:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest]
+        python-version: [2.7, 3.7]
+    env:
+      OS: ${{ matrix.os }}
+      PYTHON: ${{ matrix.python-version }}
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v2
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install pytest pytest-cov coveralls
+        pip install -r requirements.txt
+    - name: Install dorado
+      run: |
+        pip install -e .
+    - name: Run examples
+      run: |
+        python examples/steady_anuga_particles.py
+        python examples/steady_deltarcm_particles.py
+        python examples/steepest_descent_deltarcm.py
+        python examples/timing_anuga_particles.py
+        python examples/set_timestep_anuga_particles.py
+        python examples/draw_anuga_particle_paths.py
+        python examples/draw_deltarcm_particle_paths.py
+        python examples/true_random_walk.py
+        python examples/parallel_comparison.py
+        python examples/traveltime_straight_channel.py
+        cd examples
+        python unsteady_example.py
+
+
   docs:
     runs-on: ubuntu-latest
     steps:
