### 2020-10-02 3c631a4
only test examples on python v3.7
