### 2020-10-02 passaH2O/dorado#10 @gassmoeller
Very nice, thanks! This issue can be closed when #14 is merged.
