### 2020-10-02 passaH2O/dorado#12 @gassmoeller
Yes the suggested names sound very reasonable. You may want to change `LagrangianWalker.py` into `lagrangian_walker.py` to stay consistent with your naming and capitalization scheme, unless you had a specifc reason to make it CamelCase.
