+++ dorado/__init__.py
@@ -1,6 +1,6 @@
-__version__ = "1.0.0"
+__version__ = "2.0.0"
 
-from . import particle_tools
+from . import lagrangian_walker
 from . import parallel_routing
 from . import particle_track
 from . import routines
