### 2020-10-02 669a48e
remove Tools class, just use functions in lagrangian_walker.py
