+++ dorado/lagrangian_walker.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-Particle tools to manage the internal functions related to the routing.
+Core functions to handle the Lagrangian random walk movement of the particles.
 
 Project Homepage: https://github.com/passaH2O/dorado
 """
@@ -11,380 +11,376 @@ import numpy as np
 from numpy.random import random
 
 
-class Tools():
-    """Class to hold the internal functions listed below.
+def random_pick_seed(choices, probs=None):
+    """Randomly pick a number from array of choices.
 
-        **random_pick_seed** :
-            Random draw for initial particle location given a set of potential
-            locations or a region to seed particles
+    **Inputs** :
 
-        **get_weight** :
-            Pull the weights for the random walk for surrounding cells and
-            choose the new cell location
+        choices : `ndarray`
+            Array of possible values to draw from
 
-        **calculate_new_ind** :
-            Calculate the new particle index locations
+        probs : `ndarray`
+            *Optional*, can add weighted probabilities to draw
 
-        **step_update** :
-            Checking that the new indices are in fact some distance away from
-            the previous particle location
+    **Outputs** :
 
-        **calc_travel_times** :
-            Calculate the particle travel time to make the step from the
-            previous location to the new one by using the inverse of the
-            average velocity (averaged between the previous and new cell
-            locations)
-
-        **check_for_boundary** :
-            Function to check and disallow particle to travel outside of the
-            domain
-
-        **random_pick** :
-            Function to do the weighted random draw for the walk
-
-    The *Tools* class is inherited by the *Particle* class which contains the
-    broader functions initializing the particle parameters (*__init__*) and to
-    do a single iteration of the particle transport (**run_iteration**)
+        choices[idx] : `int`
+            The randomly chosen value
 
     """
+    # randomly pick tracer drop cell to use given a list of potential spots
+    if not probs:
+        probs = np.array([1 for i in list(range(len(choices)))])
+    # find the corresp. index value from input 'choices' list of indices
+    cutoffs = np.cumsum(probs)
+    idx = cutoffs.searchsorted(np.random.uniform(0, cutoffs[-1]))
 
-    # random pick seeding location
-    def random_pick_seed(self, choices, probs=None):
-        """Randomly pick a number from array of choices.
+    return choices[idx]
 
-        **Inputs** :
 
-            choices : `ndarray`
-                Array of possible values to draw from
+def get_weight(Particles, ind):
+    """Assign weights to cells surrounding current index.
 
-            probs : `ndarray`
-                *Optional*, can add weighted probabilities to draw
+    Function to assign weights to the surrounding 8 cells around the
+    current index and randomly choose one of those cells.
 
-        **Outputs** :
+    **Inputs** :
 
-            choices[idx] : `int`
-                The randomly chosen value
+        Particles : :obj:`Particles`
+            A :obj:`dorado.Particles` object
 
-        """
-        # randomly pick tracer drop cell to use given a list of potential spots
-        if not probs:
-            probs = np.array([1 for i in list(range(len(choices)))])
-        # find the corresp. index value from input 'choices' list of indices
-        cutoffs = np.cumsum(probs)
-        idx = cutoffs.searchsorted(np.random.uniform(0, cutoffs[-1]))
+        ind : `tuple`
+            Tuple (x,y) with the current location indices
 
-        return choices[idx]
+    **Outputs** :
 
-    # pull weights
-    def get_weight(self, ind):
-        """Assign weights to cells surrounding current index.
+        new_cell : `int`
[…]
