### 2020-10-03 1fb0f25
make tests match new code structure - actual contents unchanged
