+++ tests/test_particle_track.py
@@ -24,7 +24,7 @@ params.qx = np.zeros((3,3))
 params.qy = np.ones((3,3))
 params.theta = 1
 params.model = 'DeltaRCM'
-particle = particle_track.Particle(params)
+particle = particle_track.Particles(params)
 
 # testing of the Particle __init__ functionality
 def test_xseed():
@@ -60,12 +60,12 @@ def test_theta():
 
 def test_neg_diffcoeff():
     params.diff_coeff = -1.0
-    particle = particle_track.Particle(params)
+    particle = particle_track.Particles(params)
     assert particle.diff_coeff == 0.0
 
 def test_big_diffcoeff():
     params.diff_coeff = 3.0
-    particle = particle_track.Particle(params)
+    particle = particle_track.Particles(params)
     assert particle.diff_coeff == 3.0
 
 def test_dry_depth():
@@ -102,12 +102,12 @@ def test_steep():
 def test_steep_true():
     params.diff_coeff = 'invalid'
     params.steepest_descent = True
-    particle = particle_track.Particle(params)
+    particle = particle_track.Particles(params)
     assert particle.steepest_descent == True
 
 def test_steep_other():
     params.steepest_descent = 'other'
-    particle = particle_track.Particle(params)
+    particle = particle_track.Particles(params)
     assert particle.steepest_descent == False
 
 # testing of the run_iteration function
@@ -147,7 +147,7 @@ def test_travel_time_given():
 
 def test_init_params():
     # test initialization of params class
-    params = particle_track.params()
+    params = particle_track.modelParams()
     # make assertions
     assert params.seed_xloc is None
     assert params.seed_yloc is None
@@ -179,24 +179,24 @@ class TestValueErrors:
     """
 
     def test_seedx(self):
-        params = particle_track.params()
+        params = particle_track.modelParams()
         # expect ValueError
         with pytest.raises(ValueError):
-            particle = particle_track.Particle(params)
+            particle = particle_track.Particles(params)
 
     def test_seedy(self):
-        params = particle_track.params()
+        params = particle_track.modelParams()
         params.seed_xloc = [1]
         # expect ValueError
         with pytest.raises(ValueError):
-            particle = particle_track.Particle(params)
+            particle = particle_track.Particles(params)
 
     def test_seedx_none(self):
         class pobj():
             pass
         params = pobj
         with pytest.raises(ValueError):
-            particle = particle_track.Particle(params)
+            particle = particle_track.Particles(params)
 
     def test_seedy_none(self):
         class pobj():
@@ -204,44 +204,44 @@ class TestValueErrors:
         params = pobj
         params.seed_xloc = [1]
         with pytest.raises(ValueError):
-            particle = particle_track.Particle(params)
+            particle = particle_track.Particles(params)
 
     def test_num_tracers(self):
-        params = particle_track.params()
+        params = particle_track.modelParams()
         params.seed_xloc = [1]
         params.seed_yloc = [1]
         with pytest.raises(ValueError):
-            particle = particle_track.Particle(params)
+            particle = particle_track.Particles(params)
 
     def test_dx(self):
-        params = particle_track.params()
+        params = particle_track.modelParams()
         params.seed_xloc = [1]
         params.seed_yloc = [1]
         params.Np_tracer = 1
         with pytest.raises(ValueError):
-            particle = particle_track.Particle(params)
+            particle = particle_track.Particles(params)
 
     def test_broken_depth(self):
-        params = particle_track.params()
+        params = particle_track.modelParams()
         params.seed_xloc = [1]
         params.seed_yloc = [1]
         params.Np_tracer = 1
         params.dx = 1
[…]
