### 2020-10-03 571f636
breakout generator function, but retain ability to provide particle info to modelParams
