+++ dorado/particle_track.py
@@ -26,17 +26,6 @@ class modelParams:
 
     **Required Parameters:**
 
-        seed_xloc : `list`
-            List of x-coordinates over which to initially distribute the
-            particles
-
-        seed_yloc : `list`
-            List of y-coordinates over which to initially distribute the
-            particles
-
-        Np_tracer : `int`
-            Number of particles to use
-
         dx : `float`
             Length along one square cell face
 
@@ -104,6 +93,17 @@ class modelParams:
             weighted random walk. If True, then the highest weighted cells are
             used to route the particles. Default value is False.
 
+        seed_xloc : `list`
+            List of x-coordinates over which to initially distribute the
+            particles
+
+        seed_yloc : `list`
+            List of y-coordinates over which to initially distribute the
+            particles
+
+        Np_tracer : `int`
+            Number of particles to use
+
     This list of expected parameter values can also be obtained by querying the
     class attributes with `dir(modelParams)`, `modelParams.__dict__`, or
     `vars(modelParams)`.
@@ -119,9 +119,6 @@ class modelParams:
         handled on a case-by-case basis.
 
         """
-        self.seed_xloc = None
-        self.seed_yloc = None
-        self.Np_tracer = None
         self.dx = None
         self.depth = None
         self.stage = None
@@ -129,6 +126,10 @@ class modelParams:
         self.qy = None
         self.u = None
         self.v = None
+        # initialize Np_tracer, seed_xloc, and seed_yloc in case
+        self.Np_tracer = None
+        self.seed_xloc = None
+        self.seed_yloc = None
 
 
 class Particles():
@@ -146,26 +147,6 @@ class Particles():
 
         """
         # REQUIRED PARAMETERS #
-        # Define the seeding locations as list of x and y locations
-        if getattr(params, 'seed_xloc', None) is None:
-            raise ValueError("No tracer seeding x-locations"
-                             " (modelParams.seed_xloc) have been defined")
-        else:
-            self.seed_xloc = params.seed_xloc
-
-        if getattr(params, 'seed_yloc', None) is None:
-            raise ValueError("No tracer seeding y-locations"
-                             " (modelParams.seed_yloc) have been defined")
-        else:
-            self.seed_yloc = params.seed_yloc
-
-        # Define the number of tracers to be simulated
-        if getattr(params, 'Np_tracer', None) is None:
-            raise ValueError("Number of tracer particles"
-                             " (modelParams.Np_tracer) has not been defined")
-        else:
-            self.Np_tracer = params.Np_tracer
-
         # Define the length along one cell face (assuming square cells)
         if getattr(params, 'dx', None) is None:
             raise ValueError("Length of cell face (modelParams.dx) is"
@@ -405,36 +386,45 @@ class Particles():
                                     [pi, 0, 0],
                                     [5*pi/4, 3*pi/2, 7*pi/4]])
 
-    # run an iteration where particles are moved
-    # have option of specifying the particle start locations
-    # otherwise they are randomly placed within x and y seed locations
-    def run_iteration(self,
-                      start_xindices=None,
-                      start_yindices=None,
-                      start_times=None,
-                      previous_walk_data=None,
-                      target_time=None):
-        """Run an iteration of the particle routing.
+        # load in Np_tracer, seed_xloc and seed_yloc if they exist
+        if getattr(params, 'seed_xloc', None) is not None:
+            self.seed_xloc = params.seed_xloc
 
-        Runs an iteration of the particle routing.
-        Returns at each step the particle's locations and travel times.
+        if getattr(params, 'seed_yloc', None) is not None:
+            self.seed_yloc = params.seed_yloc
+
+        if getattr(params, 'Np_tracer', None) is not None:
[…]
