+++ tests/test_particle_track.py
@@ -112,38 +112,34 @@ def test_steep_other():
 
 # testing of the run_iteration function
 def test_start_pairs_X():
-    all_walk_data = particle.run_iteration()
+    init_walk_data = particle.generate_particles()
+    all_walk_data = particle.run_iteration(init_walk_data)
     assert all_walk_data['xinds'][0][0] == params.seed_xloc[0]
 
 def test_start_pairs_Y():
-    all_walk_data = particle.run_iteration()
+    init_walk_data = particle.generate_particles()
+    all_walk_data = particle.run_iteration(init_walk_data)
     assert all_walk_data['yinds'][0][0] == params.seed_yloc[0]
 
 def test_start_pairs_X2():
-    all_walk_data = particle.run_iteration(start_xindices=params.seed_xloc)
+    init_walk_data = particle.generate_particles(seed_xloc=params.seed_xloc)
+    all_walk_data = particle.run_iteration(init_walk_data)
     assert all_walk_data['xinds'][0][0] == params.seed_xloc[0]
 
 def test_start_pairs_Y2():
-    all_walk_data = particle.run_iteration(start_yindices=params.seed_yloc)
+    init_walk_data = particle.generate_particles(seed_yloc=params.seed_yloc)
+    all_walk_data = particle.run_iteration(init_walk_data)
     assert all_walk_data['yinds'][0][0] == params.seed_yloc[0]
 
 def test_travel_time():
     # Particle doesn't travel in the 3x3 space due to the 'sticky' edge
     # conditions so check that travel time is 0 and particle hasn't moved
     np.random.seed(0)
-    all_walk_data = particle.run_iteration()
+    init_walk_data = particle.generate_particles()
+    all_walk_data = particle.run_iteration(init_walk_data)
     assert all_walk_data['xinds'][0][0] == 1
     assert all_walk_data['yinds'][0][0] == 1
     assert all_walk_data['travel_times'][0][0] == 0.0
-
-
-def test_travel_time_given():
-    # Particle doesn't travel in the 3x3 space due to the 'sticky' edge
-    # conditions so check that travel time is 0 and particle hasn't moved
-    np.random.seed(0)
-    all_walk_data = particle.run_iteration(start_times=[0.0])
-    assert pytest.approx(all_walk_data['travel_times'][0][0] == 0.0)
-
 
 def test_init_params():
     # test initialization of params class
@@ -164,10 +160,11 @@ def test_init_params():
 def test_previous_walk_data():
     # test of loading previously defined walk data
     np.random.seed(0)
-    old_walk_data = particle.run_iteration()
+    old_init = particle.generate_particles()
+    old_walk_data = particle.run_iteration(old_init)
     # try to do another walk - test just makes sure code doesn't break
     np.random.seed(0)
-    all_walk_data = particle.run_iteration(previous_walk_data=old_walk_data)
+    all_walk_data = particle.run_iteration(init_walk_data=old_walk_data)
     assert all_walk_data['xinds'][0][0] == 1
     assert all_walk_data['yinds'][0][0] == 1
     assert all_walk_data['travel_times'][0][0] == 0.0

[…]
