### 2020-10-06 0428a02
remove particle info from params, 2 methods for particle seeding, random and exact
