### 2020-10-06 e327f93
refactor walk_data to be an attribute of Particles
