+++ dorado/particle_track.py
@@ -374,6 +374,14 @@ class Particles():
         # initialize number of particles as 0
         self.Np_tracer = 0
 
+        # initialize the walk_data
+        self.walk_data = None
+
+    # function to clear walk data if you've made a mistake while generating it
+    def clear_walk_data(self):
+        """Manually reset self.walk_data back to None."""
+        self.walk_data = None
+
     # generate a group of particles in a set of initial locations
     def generate_particles(self,
                            Np_tracer,
@@ -388,6 +396,11 @@ class Particles():
         particles are to be seeded in different groupings in different
         regions, this function can be called multiple times in succession
         prior to moving them via :obj:`run_iteration`.
+
+        Walk data is stored within :obj:`Particles` as `Particles.walk_data`
+        so if this method is called in succession, even without the
+        `previous_walk_data` flag, the new particle seed locations will be
+        appended to the `walk_data` attribute of :obj:`Particles`.
 
         **Inputs** :
 
@@ -410,15 +423,13 @@ class Particles():
 
             previous_walk_data : `dict`, optional
                 Dictionary of all prior x locations, y locations, and travel
-                times. Order of indices is
+                times. This input parameter should only be used if 'new' or
+                'indpendent' walk data exists (e.g. data loaded from a .json
+                file). Otherwise the walk data stored in `self.walk_data` is
+                likely to contain the previous information. Order of indices is
                 previous_walk_data[field][particle][iter], where e.g.
                 ['travel_times'][5][10] is the travel time of the 5th particle
                 at the 10th iteration.
-
-        **Outputs** :
-
-            init_walk_data : `dict`
-                Dictionary of initial x and y locations and travel times.
 
         """
         # if the values in self are invalid the input checked will catch them
@@ -453,8 +464,8 @@ class Particles():
                     if Np_to_seed <= 0:
                         break
                     # otherwise keep on seeding
-                    new_start_xindices = new_start_xindices + seed_xloc[i]
-                    new_start_yindices = new_start_yindices + seed_yloc[i]
+                    new_start_xindices = new_start_xindices + [seed_xloc[i]]
+                    new_start_yindices = new_start_yindices + [seed_yloc[i]]
                     Np_to_seed -= 1  # reduce number of particles left to seed
 
         # Now initialize vectors that will create the structured list
@@ -463,6 +474,11 @@ class Particles():
         new_yinds = [[new_start_yindices[i]] for i in
                      list(range(Np_tracer))]
         new_times = [[new_start_times[i]] for i in list(range(Np_tracer))]
+
+        # establish the start values
+        start_xindices = new_xinds
+        start_yindices = new_yinds
+        start_times = new_times
 
         if previous_walk_data is not None:
             # If the generator has been run before, or if new
@@ -474,14 +490,22 @@ class Particles():
             prev_times = previous_walk_data['travel_times']
 
             # combine previous and new lists of particle information
-            start_xindices = prev_xinds + new_xinds
-            start_yindices = prev_yinds + new_yinds
-            start_times = prev_times + new_times
+            start_xindices = prev_xinds + start_xindices
+            start_yindices = prev_yinds + start_yindices
+            start_times = prev_times + start_times
 
-        else:
-            start_xindices = new_xinds
-            start_yindices = new_yinds
-            start_times = new_times
+        if self.walk_data is not None:
+            # if there is walk_data from a previous call to the generator,
+            # or from simulating particle transport previously, then we want
+            # to keep it and append the new data to it
+            internal_xinds = self.walk_data['xinds']
[…]
