### 2020-10-06 0af71d4
update unit tests for new structure
