### 2020-10-06 42e8c67
update the examples for new code structure
