+++ examples/steady_anuga_particles.py
@@ -23,15 +23,20 @@ params.stage = depth  # for this example we don't have stage data
 params.depth = depth
 params.qx = qx
 params.qy = qy
-
-params.seed_xloc = list(range(20, 30))
-params.seed_yloc = list(range(48, 53))
-params.Np_tracer = 50
 params.dx = 50.
 params.model = 'Anuga'
+
+# define initial particle locations and number of particles
+seed_xloc = list(range(20, 30))
+seed_yloc = list(range(48, 53))
+Np_tracer = 50
 
 # Apply the parameters to run the particle routing model
 np.random.seed(0)  # fix the random seed for example
 
+# initialize particles and generate particles
+particles = pt.Particles(params)
+particles.generate_particles(Np_tracer, seed_xloc, seed_yloc)
+
 # using steady (time-invariant) plotting routine
-walk_data = steady_plots(params, 50, 'steady_anuga_example')
+walk_data = steady_plots(particles, 50, 'steady_anuga_example')
