### 2020-10-06 7e53acd
fix time_plots routine and example
