### 2020-10-06 88e8b6d
bringing documentation up to speed
