+++ examples/timing_anuga_particles.py
@@ -38,4 +38,9 @@ particles.generate_particles(Np_tracer, seed_xloc, seed_yloc)
 
 # Apply the parameters to run the particle routing model
 # using steady (time-invariant) plotting routine
-time_plots(particles, 50, 'timing_anuga_example')
+walk_data = time_plots(particles, 50, 'timing_anuga_example')
+
+from dorado.routines import get_state
+_, _, finaltimes = get_state(walk_data)
+print('List of particle travel times for final particle locations: ' +
+      str(np.round(finaltimes)))
