+++ docs/source/examples/example10.rst
@@ -11,11 +11,13 @@ Full example script available :download:`here <../../../examples/parallel_compar
 
 We will be running the parallel comparison on the ANUGA example domain used in :ref:`example01`. The only difference is that we will use a larger number of particles so that the impact of the parallelization can be measured even on this relatively small and simple domain.
 
-So after initializing the parameters as we did in :ref:`example01`, we will change the number of particles being used.
+So after initializing the parameters as we did in :ref:`example01`, we will stick some of the particle parameters into the `params` object for convenience.
 
 .. doctest::
 
    >>> params.Np_tracer = 200
+   >>> params.seed_xloc = list(range(20, 30))
+   >>> params.seed_yloc = list(range(48, 53))
 
 Next we will do the parallel routing using 2 CPU cores using the `parallel_routing` function. In this case, we are routing 200 particles on each core, and iterating each particle 50 times.
 
@@ -31,8 +33,10 @@ For comparison, we will the comparable serial process twice so that the proper n
    >>> for z in list(range(0, 2)):
    >>>    all_walk = None  # initialize walk data list
    >>>    particle = Particle(params)
+   >>>    particle.generate_particles(params.Np_tracer, params.seed_xloc)
+   >>>                                params.seed_yloc
    >>>    for i in list(range(0, 50)):
-   >>>        all_walk = particle.run_iteration(previous_walk_data=all_walk)
+   >>>        all_walk = particle.run_iteration()
 
 If we were to time this example (timing included in full example script), a roughly 2x speed-up can be observed. Speed-up may vary based on a variety of factors.
 

[…]
