+++ _downloads/e29693d41e132dbe5f99a7109436322c/timing_anuga_particles.py
@@ -1,11 +1,14 @@
 """Example of the workflow with gridded anuga output data"""
 
 import numpy as np
+import os.path
 from dorado.routines import time_plots
 from dorado.particle_track import params
 
-# load some variables from a deltarcm output so stage is varied
-data = np.load('ex_anuga_data.npz')
+# load some variables from an anuga output so stage is varied
+f_path = os.path.abspath(os.path.dirname(__file__))
+data_path = os.path.join(f_path, 'ex_anuga_data.npz')
+data = np.load(data_path)
 
 # pull depth and stage from that data
 depth = data['depth']
@@ -21,8 +24,8 @@ params.stage = depth  # using depth as stand-in for stage in this example
 params.qx = qx
 params.qy = qy
 
-params.seed_xloc = list(range(20,30))
-params.seed_yloc = list(range(48,53))
+params.seed_xloc = list(range(20, 30))
+params.seed_yloc = list(range(48, 53))
 params.Np_tracer = 50
 params.dx = 50.
 params.theta = 1.0
