### 2020-10-06 812e907
fix runtime warning by pointing stage and depth at different places in memory
