+++ examples/timing_anuga_particles.py
@@ -20,7 +20,7 @@ params = pt.modelParams()
 
 # define the params variables
 params.depth = depth
-params.stage = depth  # using depth as stand-in for stage in this example
+params.stage = np.copy(depth)  # depth as stand-in for stage in this example
 params.qx = qx
 params.qy = qy
 params.dx = 50.
