### 2020-10-06 passaH2O/dorado#11 @elbeejay
Hi @gassmoeller, thanks for bringing this to our attention.

This ended up being due to the simplistic nature of our example cases in which we made the mistake of pointing our variables `stage` and `depth` to the same array in memory. This resulted in `nan` values entering the depth array, which later raised the warnings you pointed out.

In #15, we resolve this by updating the examples to use a copy of the array for setting one of the stage/depth variables and this warnings go away as the `depth` and `stage` variables no longer point to the same object in memory.
