### 2020-10-07 cd9e5f3
update parallel method and example
