### 2020-10-07 643ecc8
particle generation walk_data order is now new - self.walk_data - previous_walk_data
