### 2020-10-07 64959c5
amend check on types for the particle generation
