### 2020-10-07 0ff4f6d
clean up docstring
