+++ tests/test_examplecases.py
@@ -32,8 +32,7 @@ def test_few_steps_RCM():
     # defining / initializing
     particle = Particles(params) # define the particle
     np.random.seed(0) # fix the random seed
-    all_walk_data = particle.generate_particles(Np_tracer, seed_xloc,
-                                                seed_yloc) # init the walk data
+    particle.generate_particles(Np_tracer, seed_xloc, seed_yloc) # init the walk data
 
     # 3 iterations
     for i in list(range(0,3)):
@@ -62,8 +61,7 @@ def test_set_time_RCM_previousdata():
     np.random.seed(0) # fix the random seed
 
     # generate the particles
-    all_walk_data = particle.generate_particles(Np_tracer, seed_xloc,
-                                                seed_yloc) # init the walk data
+    particle.generate_particles(Np_tracer, seed_xloc, seed_yloc) # init the walk data
 
     # set time
     all_walk_data = particle.run_iteration(target_time=5e6)
@@ -91,8 +89,7 @@ def test_set_time_RCM():
     np.random.seed(0) # fix the random seed
 
     # generate particles
-    init_walk_data = particle.generate_particles(Np_tracer, seed_xloc,
-                                                seed_yloc) # init the walk data
+    particle.generate_particles(Np_tracer, seed_xloc, seed_yloc) # init the walk data
 
     # set time
     all_walk_data = particle.run_iteration(target_time=5e6)
@@ -134,8 +131,7 @@ def test_few_steps_anuga():
     # defining / initializing
     an_particle = Particles(an_params) # define the particle
     np.random.seed(0) # fix the random seed
-    all_walk_data = an_particle.generate_particles(Np_tracer, seed_xloc,
-                                                seed_yloc) # init the walk data
+    an_particle.generate_particles(Np_tracer, seed_xloc, seed_yloc) # init the walk data
 
     # 3 iterations
     for i in list(range(0,3)):
@@ -162,8 +158,7 @@ def test_boundary_anuga():
     # defining / initializing
     an_particle = Particles(an_params) # define the particle
     np.random.seed(0) # fix the random seed
-    all_walk_data = an_particle.generate_particles(Np_tracer, seed_xloc,
-                                                seed_yloc) # init the walk data
+    an_particle.generate_particles(Np_tracer, seed_xloc, seed_yloc) # init the walk data
 
     # 20 iterations
     for i in list(range(0, 20)):
@@ -190,8 +185,7 @@ def test_boundary_travel_time_anuga():
     np.random.seed(0) # fix the random seed
 
     # generate particles
-    init_walk_data = an_particle.generate_particles(Np_tracer, seed_xloc,
-                                                seed_yloc) # init the walk data
+    an_particle.generate_particles(Np_tracer, seed_xloc, seed_yloc) # init the walk data
 
     # set target time for iterations
     all_walk_data = an_particle.run_iteration(target_time=1000.0)
