### 2020-10-08 502315d
minor rewording, clean up tests
