### 2020-10-08 02be73c
drop particle info from parameter tests
