### 2020-10-09 9d541fc
fix unstructured example and update paper
