+++ joss/paper.md
@@ -24,33 +24,42 @@ bibliography: paper.bib
 
 # Summary
 
-Hydrodynamic simulations of flow through different landscapes are used to answer questions related to the transport of water, nutrients, pollutants, biota, and sediment through waterways [@Duan2006; @Rynne2016; @Wild-Allen2016; @Lauzon2018; @Czuba2019]. In geophysical systems, such as rivers, estuaries, and deltas, hydrodynamic models typically solve the depth-integrated ``shallow water" equations in an Eulerian reference frame, which is concerned with fluxes through a given region of space -- examples of these solvers include ANUGA [@anuga], Delft3D [@delft3d], Frehd [@Hodges2014] and others. However, the spatial and temporal characteristics of the movement of material through a landscape are often better understood using a Lagrangian reference frame [@Doyle2009], which follows the movement of individual objects or parcels. In this paper, we present an open-source Python package, *dorado*, which provides a transparent and accessible method for researchers to simulate passive Lagrangian particle transport on top of Eulerian hydrodynamic solutions. This mixed Eulerian-Lagrangian methodology adapts the routing functionality from the popular numerical model DeltaRCM [@Liang2015a; @Liang2015] for use with the outputs of any shallow-water hydrodynamic solver.
+Hydrodynamic simulations of flow through landscapes allow scientists to answer questions related to the transport of water, nutrients, pollutants, biota, and sediment through waterways [@Duan2006; @Rynne2016; @Wild-Allen2016; @Lauzon2018; @Czuba2019]. In geophysical systems, such as rivers, estuaries, and deltas, hydrodynamic models typically solve the depth-integrated ``shallow water" equations in an Eulerian reference frame, which is concerned with fluxes through a given region of space -- examples of these solvers include ANUGA [@anuga], Delft3D [@delft3d], Frehd [@Hodges2014] and others. However, the spatial and temporal characteristics of the movement of material through a landscape are often better understood using a Lagrangian reference frame [@Doyle2009], which follows the movement of individual objects or parcels. In this paper, we present an open-source Python package, *dorado*, which provides a transparent and accessible method for researchers to simulate passive Lagrangian particle transport on top of Eulerian hydrodynamic solutions. This mixed Eulerian-Lagrangian methodology adapts the routing functionality from the popular numerical model DeltaRCM [@Liang2015a; @Liang2015] for use with the outputs of any shallow-water hydrodynamic solver.
 
 # Statement of Need
 
-Existing particle tracking software typically contains complicated code structures [@freshkiss3d; @delft3d] with steep learning-curves, are proprietary [@fluent], or were developed to solve problems at very different spatial scales [@Yeung1989; @Dagestad2018] than typical riverine applications. This package fills a gap in the available methods for Lagrangian particle simulation by providing a flexible, open, and transparent framework which can be used in conjunction with any landscape-2D hydrodynamic model. In addition, *dorado* comes with builtin pre-processing, analysis, and plotting functionality, including methods to compute the exposure time distribution of particles to specific sub-regions of the domain, which is often of interest in ecological applications [@Kadlec2008].
[…]
