+++ docs/source/quickstart/demo01.rst
@@ -13,9 +13,9 @@ First we load and create the sample particles object.
 
 .. doctest::
 
-    >>> from dorado.example_data import define_params as dp
-    >>> import dorado as pr
+    >>> import dorado
     >>> import matplotlib.pyplot as plt
+    >>> from dorado.example_data import define_params as dp
     >>> rcmparticles = dp.make_rcm_particles()
 
 We can visualize the water depth for this scenario from the `rcmparticles` object. If you'd like to download this portion of the demo as a standalone script, it is available :download:`here <../pyplots/quickstart/demo1_depth.py>`.
@@ -33,7 +33,7 @@ Now let's route 50 particles for 50 iterations.
 
 .. doctest::
 
-    >>> pr.routines.steady_plots(rcmparticles, 50, 'demo-1')
+    >>> dorado.routines.steady_plots(rcmparticles, 50, 'demo-1')
 
 The `steady_plots()` function saves plots of each iteration of the particle movement to the subfolder 'demo-1/figs'. The final plot is shown below; the initial particle locations are shown as blue dots, and the final particle locations are red dots. If you'd like to download this portion of the demo as a standalone script, it is available :download:`here <../pyplots/quickstart/demo1.py>`.
 
