+++ docs/source/examples/example06.rst
@@ -28,8 +28,8 @@ To set a travel time target, we will use the lower-level API and access the `par
 .. doctest::
 
    >>> walk_data = particle.run_iteration(target_time=2100)
-   >>> pr.routines.plot_state(anugaparams.depth, walk_data, iteration=0, c='b')
-   >>> pr.routines.plot_final(anugaparams.depth, walk_data, iteration=-1, c='r')
+   >>> routines.plot_state(params.depth, walk_data, iteration=0, c='b')
+   >>> routines.plot_final(params.depth, walk_data, iteration=-1, c='r')
    >>> plt.title('Initial and Final Particle Locations')
    >>> plt.show()
 
@@ -39,7 +39,7 @@ To see how close we are to the prescribed target travel time of 2100 seconds, we
 
 .. doctest::
 
-   >>> _, _, finaltimes = get_state(walk_data)
+   >>> _, _, finaltimes = routines.get_state(walk_data)
    >>> print('List of particle travel times for final particle locations: ' +
    >>>       str(np.round(finaltimes)))
 
