+++ docs/source/examples/example11.rst
@@ -14,10 +14,12 @@ Shortened code for the synthetic channel creation will be provided, for the full
    >>> import dorado.particle_track as pt
 
    >>> domain = np.zeros((100, 50))
+   >>> depth = np.zeros_like(domain)
    >>> depth[:, 10:40] = 1.0
+   >>> v = np.zeros_like(domain)
    >>> v[:, 10:40] = -10.0
 
-In the 100 x 50 cell domain, we have a straight channel with flow from the top of the domain to the bottom at -10 m/s. Our cell size is 50m, and the particles will be seeded at row 10. We will define a region of interest that extends to row 70, and apply the :obj:`dorado.particle_track.exposure_time` function to calculate the amount of times the particles have traveled. The region will end at the boundary of row 70. Since the flow velocity is both steady and uniform in this channel, our expected travel time is simply:
+In the 100 x 50 cell domain, we have a straight channel with flow from the top of the domain to the bottom at -10 m/s. Our cell size is 50m, and the particles will be seeded at row 10. We will define a region of interest that extends to row 70, and apply the :obj:`dorado.particle_track.exposure_time` function to calculate the travel times of the particles. The region will end at the boundary of row 70. Since the flow velocity is both steady and uniform in this channel, our expected travel time is simply:
 
 .. math::
    :nowrap:
@@ -32,7 +34,7 @@ Shortened parameter initialization:
 
 .. doctest::
 
-   >>> params = pt.params()
+   >>> params = pt.modelParams()
    >>> params.depth = depth
    >>> params.v = v
    >>> params.dx = dx
@@ -42,7 +44,7 @@ Then we will define `diff_coeff`, initialize the Particle class, and do the simu
 
 .. doctest::
 
-   >>> for dc in range(0, 2):
+   >>> for dc in list(range(0, 2)):
    >>>    # set diff_coeff
    >>>    if dc == 0:
    >>>        params.diff_coeff = 0.0
@@ -50,7 +52,7 @@ Then we will define `diff_coeff`, initialize the Particle class, and do the simu
    >>>        params.diff_coeff = 1.0
 
    >>>    # make particle
-   >>>    particle = pt.Particle(params)
+   >>>    particle = pt.Particles(params)
 
    >>>    # walk it
    >>>    particle.generate_particles(Np_tracer, seed_xloc, seed_yloc)
