### 2020-10-13 #2585 @kbarnhart
@elbeejay it looks like you and co-authors are making great progress addressing the reviewer comments in the source repository. I wanted to check in and see if you had any questions for me.

At this stage of the JOSS review process, we (editor, reviewers) typically just wait for the submitting authors to address any outstanding issues and provided feedback as needed. It looks like you and @gassmoeller are already having this sort of discussion in your in-repo issues. 🎉

There is no standard timeframe for this stage of the review process and we typically don't "pause" the JOSS review (to answer the question raised [here](https://github.com/passaH2O/dorado/issues/13#issuecomment-702865879).
