+++ joss/paper.md
@@ -36,16 +36,16 @@ Existing software for tracking water parcels or particles in a Lagrangian framew
 \begin{equation}
     F^{*} = \gamma F_{sfc} + (1-\gamma) F_{int}
 \end{equation}
-in which $F_{sfc}$ and $F_{int}$ represent the surface gradient and inertial components, respectively. Modifying $\gamma$ indirectly controls the diffusivity of the travel path. The second weighting parameter, $\theta$, modifies the routing weights for the random walk based on the local water depth value, $h_i$. The routing weights for each neighboring cell are calculated per [@Liang2015a]:
+in which $F_{sfc}$ and $F_{int}$ represent the surface gradient and inertial components, respectively. Modifying $\gamma$ indirectly controls the diffusivity of the travel path. The second weighting parameter, $\theta$, modifies the routing weights, $w_i$, for the random walk based on the local water depth value, $h_i$. The routing weights for each neighboring cell are calculated per @Liang2015a:
 
 \begin{equation}
     w_i = \frac{h_i^\theta \max \left(0, F \cdot d_i \right)}{\Delta_i}
 \end{equation}
-wherein $w_i$ are the routing weights, $F$ are the routing directions, $d_i$ represents the unit vector from the current cell to its neighbor, and $\Delta_i$ is the D8 cellular distance. Here, the local flow depth acts as a resistance term -- as $\theta$ gets larger, the local routing weights have an increasingly large dependence on the water depth of the neighboring cells. By altering the two parameters $\gamma$ and $\theta$, the user has control over the randomness with which the particles are routed. For a full description of this methodology, see [@Liang2015; @Liang2015a] and the *dorado* documentation.
+wherein $F$ are the routing directions, $d_i$ represents the unit vector from the current cell to its neighbor, and $\Delta_i$ is the D8 cellular distance. Here, the local flow depth acts as a resistance term -- as $\theta$ gets larger, the local routing weights have an increasingly large dependence on the water depth of the neighboring cells. By altering the two parameters $\gamma$ and $\theta$, the user has control over the randomness with which the particles are routed. For a full description of this methodology, see @Liang2015, @Liang2015a and the *dorado* documentation.
 
 Particle travel times are back-calculated from their locations after routing by accounting for the flow velocity, flow direction, and grid size. The time elapsed during each particle step is assumed to be proportional to the distance traveled in the direction of the mean flow, $d_{eq}$, which is equal to the step distance $\Delta_i$ projected onto a unit vector oriented in the direction of the velocity field, $\phi$, according to $d_{eq} = \Delta_i \cdot \cos(\phi)$. During a model iteration, each particle is allowed to progress a different length of time, depending on the velocity of the flow in the vicinity of each particle and the orientation of each step in relation to the mean flow direction. The travel time of each step is then modified according to a user-specified parameter, $D_c$, which acts as a dispersion coefficient.
 
-*dorado* provides functions with which users can choose to sync up the particle evolution either by number of step iterations or by individual particle travel times. In addition, *dorado* includes several post-processing functions for computing and plotting the fraction of time particles spend ``exposed" to a region of interest, otherwise known as the exposure time distribution (or residence time distribution, in steady flows) [@Kadlec2008; @Benjamin2013; @Hiatt2018]. This travel time methodology has been tested against analytical solutions for a plug-flow reactor with dispersion [@Benjamin2013], as well as against exposure time distributions from prior models of real systems [@Hiatt2018], and performed well at reproducing the observed travel times in both systems. As with all Lagrangian methods involving finite samples of particles, we expect the travel time computations to struggle to reproduce heavy-tailed behavior observed in some systems [@Zhang2012], and expect that *dorado* will perform most accurately in advection-dominated flows.
[…]
