### 2020-10-14 36cae78
fix docs to python 3.8 for now
