+++ _downloads/88f4ed81007be3f3b1abd52282152811/true_random_walk.py
@@ -24,27 +24,31 @@ from dorado import routines
 # theta and gamma are set to 0.
 
 # create params and then assign the parameters
-params = particle_track.params()
+params = particle_track.modelParams()
 
 # define the params variables
 params.depth = np.ones((100, 100))
 params.stage = np.ones((100, 100))
 params.qx = np.zeros_like(params.depth)
 params.qy = np.zeros_like(params.depth)
-
-params.seed_xloc = list(range(45, 56))
-params.seed_yloc = list(range(45, 56))
-params.Np_tracer = 50
 params.dx = 50.
 params.theta = 0.0
 params.gamma = 0.0
 params.model = 'None'
 
-# Now we can apply the high-level API function, `steady_plots`, to simulate the
-# movement of the particles we have defined in `params` for 50 iterations.
-# Default behavior for this function is to write and save figures and particle
-# location data to the disk, for this example we are turning that option off.
+# particle info
+seed_xloc = list(range(45, 56))
+seed_yloc = list(range(45, 56))
+Np_tracer = 50
 
 np.random.seed(0)  # fix random seed for the example
+
+# initialize the particles class and generate the initial particles
+particles = particle_track.Particles(params)
+particles.generate_particles(Np_tracer, seed_xloc, seed_yloc)
+
+# Now we can apply the high-level API function, `steady_plots`, to simulate the
+# movement of the particles we have defined in `params` for 50 iterations.
+
 # using steady (time-invariant) plotting routine
-walk_data = routines.steady_plots(params, 50, 'true_random_walk')
+walk_data = routines.steady_plots(particles, 50, 'true_random_walk')
