+++ _downloads/88740e1ea0aa18ed325a9e72627b7118/particle_track.py
@@ -15,27 +15,17 @@ import numpy as np
 import scipy
 from scipy import interpolate
 from tqdm import tqdm
-from .particle_tools import Tools
+import dorado.lagrangian_walker as lw
 
 
-class params:
+class modelParams:
     """Parameter class with attributes and grid particles will be routed on.
 
-    The parameters class, `params`, must be populated with user-defined
+    The parameters class, :obj:`dorado.particle_track.modelParams`,
+    must be populated with user-defined
     attributes of the grid the particles will be modeled on.
 
     **Required Parameters:**
-
-        seed_xloc : `list`
-            List of x-coordinates over which to initially distribute the
-            particles
-
-        seed_yloc : `list`
-            List of y-coordinates over which to initially distribute the
-            particles
-
-        Np_tracer : `int`
-            Number of particles to use
 
         dx : `float`
             Length along one square cell face
@@ -105,12 +95,13 @@ class params:
             used to route the particles. Default value is False.
 
     This list of expected parameter values can also be obtained by querying the
-    class attributes with `dir(params)`, `params.__dict__`, or `vars(params)`.
+    class attributes with `dir(modelParams)`, `modelParams.__dict__`, or
+    `vars(modelParams)`.
 
     """
 
     def __init__(self):
-        """Create the expected variables for the params class.
+        """Create the expected variables for the modelParams class.
 
         Variables are initialized as NoneType they need to be assigned by the
         user. Due to the wide variety of data formats produced by differeny
@@ -118,9 +109,6 @@ class params:
         handled on a case-by-case basis.
 
         """
-        self.seed_xloc = None
-        self.seed_yloc = None
-        self.Np_tracer = None
         self.dx = None
         self.depth = None
         self.stage = None
@@ -130,14 +118,15 @@ class params:
         self.v = None
 
 
-class Particle(Tools):
+class Particles():
     """Class for the particle(s) that is(are) going to be routed."""
 
     def __init__(self, params):
         """Check input parameters and assign default values where/if needed.
 
-        Methods require a class of parameters (:obj:`params`) to be passed to
-        the Particles class. e.g. particle = Particles(params)
+        Methods require a class of parameters
+        (:obj:`dorado.particle_track.modelParams`) to be passed
+        to the Particles class. e.g. particle = Particles(modelParams)
 
         Initialization tries to assign each value from the parameter class,
         otherwise an error is raised or default values are assigned when
@@ -145,29 +134,10 @@ class Particle(Tools):
 
         """
         # REQUIRED PARAMETERS #
-        # Define the seeding locations as list of x and y locations
-        if getattr(params, 'seed_xloc', None) is None:
-            raise ValueError("No tracer seeding x-locations (params.seed_xloc)"
-                             " have been defined")
-        else:
-            self.seed_xloc = params.seed_xloc
-
-        if getattr(params, 'seed_yloc', None) is None:
-            raise ValueError("No tracer seeding y-locations (params.seed_yloc)"
-                             " have been defined")
-        else:
-            self.seed_yloc = params.seed_yloc
-
-        # Define the number of tracers to be simulated
-        if getattr(params, 'Np_tracer', None) is None:
-            raise ValueError("Number of tracer particles (params.Np_tracer)"
-                             " has not been defined")
-        else:
-            self.Np_tracer = params.Np_tracer
-
         # Define the length along one cell face (assuming square cells)
         if getattr(params, 'dx', None) is None:
-            raise ValueError("Length of cell face (params.dx) is undefined")
+            raise ValueError("Length of cell face (modelParams.dx) is"
+                             " undefined")
         else:
             self.dx = params.dx
 
[…]
