### 2020-10-14 #2585 @elbeejay
@kbarnhart I don't know where we'd update this, but some of the changes that have been made during the review process were 'breaking' changes. So in light of that, the code associated with the JOSS paper will be v2.0.0 (or maybe v2.0.x if something small has to change) instead of v1.0.0. Is there an easy way to change this? Thanks!
