### 2020-10-14 passaH2O/dorado#7 @elbeejay
Hello, we didn't want to automatically close the PR with the merge of #15 but we do believe we've addressed these concerns, and hopefully have made the paper stronger in the process. A new rendered version of the paper is available [here](https://github.com/openjournals/joss-reviews/issues/2585#issuecomment-708569220).

We've made the following changes:

1. To improve the statement of need, we try to make it more clear that "particle tracking" is analogous to the tracking of water parcels, and we state that the gap we are aiming to fill is specifically for doing this in a Lagrangian framework. To clarify the use-case and scale of problem, we state that this package is meant for "meant for rapid application in conjunction with any landscape-scale 2D hydrodynamic model"

2. We have added some of the theory behind the weighted random walk scheme into the paper "Background" section.
