### 2020-10-14 passaH2O/dorado#9 @elbeejay
We think we've addressed these concerns, and appreciate the suggestions. A point-by-point explanation of what has been done to handle each suggestion is below.

- We've dropped official support for the `conda` installation (which had it's own issues anyway) and are only recommending `pip` or installation from source. Our continuous integration ensures that all code merged into the 'master' branch installs, passes unit tests, and can run the examples on MacOS, Windows, and Linux running Python versions 2.7, 3.6, 3.7 and 3.8.

- We fixed the quickstart demo and tried to provide a bit more context as to what is happening and why someone might want to conduct a simulation like this

- While we welcome suggestions or additions to the codebase, we do not have a clear road map for future additions or functionality. But, thanks to your suggestions we have expanded upon our contribution guidelines to hopefully make contributing easier and more accessible to all.

- The primary reason we maintain Python 2.7 compatibility is so that this package can be used in the same environment as [Anuga](https://github.com/GeoscienceAustralia/anuga_core), which is the basis for many of our example cases, and is the hydrodynamic model some of our users are working with. Currently Anuga is still in Python 2.7 with incomplete support for 3.x.
