### 2020-10-14 passaH2O/dorado#12 @elbeejay
We have gone through and made the proposed changes. As suggested above, we named our random walk script `lagrangian_walker.py`. The bulk of changes associated with this issue are in #15.
