### 2020-10-14 #2585 @elbeejay
@gassmoeller and @dbuscombe-usgs, we've done our best to incorporate the changes you both suggested we make to improve the quality of the paper and the code. Thank you both for the thoughtful feedback.

I've tried to address the comments directly in each issue, but didn't want to keep @-tagging you. The bulk of the changes associated with this review were made in PR [15](https://github.com/passaH2O/dorado/pull/15) which you can view if you prefer to see the code "diffs".

For the edits and revisions made to the paper, a new rendered version is available in the **whedon** post above.

Again, thanks to you both, feel free to comment here or in a specific issue if you feel we've not sufficiently addressed your original comment or if something is unclear.
