### 2020-10-15 #2585 @kbarnhart
@elbeejay thanks for the update and for responding to reviewer comments with changes in your submitted software.

@gassmoeller and @dbuscombe-usgs please let me know if you seen any outstanding issues or if there are remaining items that @elbeejay and co-authors should attend to.

@dbuscombe-usgs there is one check-box remaining (related to community guidelines). Since I know you are not able to check the boxes, please let me know if you think it can be checked, and I can edit it. As best I can tell the authors have addressed these concerns.

Thanks all for your contributions to this review process.
