### 2020-10-17 ffbe9d6
add and update path plotting routines
