+++ dorado/routines.py
@@ -11,6 +11,8 @@ from .particle_track import modelParams
 import numpy as np
 import matplotlib
 from matplotlib import pyplot as plt
+from matplotlib import patheffects as path_effects
+from matplotlib.collections import LineCollection
 from mpl_toolkits.axes_grid1 import make_axes_locatable
 import scipy
 import os
@@ -384,11 +386,22 @@ def get_state(walk_data, iteration=-1):
     xinds = []
     yinds = []
     times = []
+    iter_exceeds_warning = 0
     # Pull out the specified value
     for ii in list(range(Np_tracer)):
-        xinds.append(walk_data['xinds'][ii][iteration])
-        yinds.append(walk_data['yinds'][ii][iteration])
-        times.append(walk_data['travel_times'][ii][iteration])
+        if len(walk_data['xinds'][ii]) > iteration:
+            xinds.append(walk_data['xinds'][ii][iteration])
+            yinds.append(walk_data['yinds'][ii][iteration])
+            times.append(walk_data['travel_times'][ii][iteration])
+        else:
+            xinds.append(walk_data['xinds'][ii][-1])
+            yinds.append(walk_data['yinds'][ii][-1])
+            times.append(walk_data['travel_times'][ii][-1])
+            iter_exceeds_warning += 1
+    
+    # if iter_exceeds_warning > 0:
+        # print('Warning: %s particles have not reached %s iterations' % \
+              # (iter_exceeds_warning, iteration))
 
     return xinds, yinds, times
 
@@ -676,15 +689,20 @@ def animate_plots(start_val, end_val, folder_name):
               dpi=300)
 
 
-def draw_travel_path(depth, walk_data,
+def draw_travel_path(grid, walk_data,
                      particles_to_follow='all',
-                     output_file='travel_paths.png'):
+                     output_file='travel_paths.png',
+                     frequency=1):
     """Make a plot with the travel path of specified particles drawn out.
 
     **Inputs** :
 
-        depth : `numpy.ndarray`
-            Water depth array.
+        grid : `numpy.ndarray`
+            A 2-D grid upon which the particles will be plotted. Examples of
+            grids that might be nice to use are
+            `dorado.particle_track.modelParams.depth`,
+            `dorado.particle_track.modelParams.stage`,
+            `dorado.particle_track.modelParams.topography`.
 
         walk_data : `dict`
             Output of `steady_plots`, `unsteady_plots`, `time_plots`, as well
@@ -697,53 +715,46 @@ def draw_travel_path(depth, walk_data,
         output_file : `str`, optional
             Path to save the output image to.
 
+        frequency : `int`, optional
+            Determines how "smooth" each particle trajectory appears by
+            skipping iterations between successive points on the path.
+            Default it to show every iteration
+
     **Outputs** :
 
         Saves a plot of particle travel paths.
 
     """
-    from matplotlib import cm
-    color_index = 0
-
     if(particles_to_follow == 'all'):
         Np_tracer = len(walk_data['xinds'])
         particles_to_follow = list(range(Np_tracer))
 
     plt.figure(figsize=(7, 4), dpi=300)
-    plt.imshow(depth, cmap='bone')
-    plt.title('Particle paths overlaid on water depth')
-    cbar2 = plt.colorbar(orientation='horizontal')
-    cbar2.set_label('Water Depth [m]')
-
+    ax = plt.gca()
+    im = ax.imshow(grid, cmap='bone', alpha=0.9)
+    ax.set_title('Particle paths')
+    
+    paths = [] # Place to store particle paths
+    colors = [] # Place to store colors
     for i in tqdm(particles_to_follow):
-        # set color for this particle (using a discrete colormap)
[…]
