### 2020-10-17 efee0c2
update plotting routines
