### 2020-10-19 ccfd3dd
update examples and docs
