+++ docs/source/examples/example07.rst
@@ -19,11 +19,12 @@ First run :ref:`example02`, and then we can access and use the particle travel i
    >>> depth = data['depth']
    >>> all_walk_data = json.load(open('steady_deltarcm_example'+os.sep+'data'+os.sep+'data.txt'))
 
-Now we will apply the `draw_travel_path` function. In this example we are going to visualize the travel paths of the first 4 particles, indicated by the list we are providing to the `draw_travel_path` function.
+Now we will apply the `draw_travel_path` function. In this example we are going to visualize the travel paths of the first 4 particles, indicated by the list we are providing to the `draw_travel_path` function. We will show every 4th iteration to visually simplify the path, controlled by `interval` parameter.
 
 .. doctest::
 
    >>> draw_travel_path(depth, all_walk_data, [0, 1, 2, 3],
-   >>>                  'steady_deltarcm_example'+os.sep+'data'+os.sep+'travel_paths.png')
+   >>>           'steady_deltarcm_example'+os.sep+'figs'+os.sep+'travel_paths.png',
+   >>>           interval=2, plot_legend=True)
 
 .. image:: images/example07/travel_paths.png
