### 2020-10-19 afc96ae
fix issue with zip iterator
