### 2020-10-19 ee22a60
unlink snake_plots from particle object
