### 2020-10-20 90e6efa
update version number to 2.1.0
