+++ dorado/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.0.0"
+__version__ = "2.1.0"
 
 from . import lagrangian_walker
 from . import parallel_routing
