### 2020-10-21 passaH2O/dorado#7 @gassmoeller
Thanks, the new statement of need looks great, this can be closed from my point of view (I cant close, because it is not my issue).
