### 2020-10-21 #2585 @kbarnhart
@elbeejay FYI the next step is that I do a final copyedit of the paper and let you know if there are any necessary changes. Then we move to the final stage of the JOSS process.
