### 2020-10-21 8ec2493
minor copyedits
dont' know if it will work for liang et al a and b
