+++ joss/paper.md
@@ -32,16 +32,16 @@ Existing software for tracking water parcels or particles in a Lagrangian framew
 
 # Background
 
-*dorado* makes use of the weighted random walk framework [@Pearson1905] to model particle transport as a stochastic Markovian process, using a grid-centric approach that takes into account local water inertial components and surface slopes, in a manner modeled after the numerical model DeltaRCM [@Liang2015; @Liang2015a]. The random walk routing weights are sensitive to two user-specified parameters, $\gamma$ and $\theta$. The $\gamma$ parameter controls the proportional importance of water surface gradient and water velocity in determining the downstream routing direction $F^{*}$, according to:
+*dorado* makes use of the weighted random walk framework [@Pearson1905] to model particle transport as a stochastic Markovian process, using a grid-centric approach that takes into account local water inertial components and surface slopes, in a manner modeled after the numerical model DeltaRCM [@Liang2015a; @Liang2015b]. The random walk routing weights are sensitive to two user-specified parameters, $\gamma$ and $\theta$. The $\gamma$ parameter controls the proportional importance of water surface gradient and water velocity in determining the downstream routing direction $F^{*}$, according to:
 \begin{equation}
     F^{*} = \gamma F_{sfc} + (1-\gamma) F_{int}
 \end{equation}
-in which $F_{sfc}$ and $F_{int}$ represent the surface gradient and inertial components, respectively. Modifying $\gamma$ indirectly controls the diffusivity of the travel path. The second weighting parameter, $\theta$, modifies the routing weights, $w_i$, for the random walk based on the local water depth value, $h_i$. The routing weights for each neighboring cell are calculated per @Liang2015a:
+in which $F_{sfc}$ and $F_{int}$ represent the surface gradient and inertial components, respectively. Modifying $\gamma$ indirectly controls the diffusivity of the travel path. The second weighting parameter, $\theta$, modifies the routing weights, $w_i$, for the random walk based on the local water depth value, $h_i$. The routing weights for each neighboring cell are calculated per @Liang2015b:
 
 \begin{equation}
     w_i = \frac{h_i^\theta \max \left(0, F \cdot d_i \right)}{\Delta_i}
 \end{equation}
-wherein $F$ are the routing directions, $d_i$ represents the unit vector from the current cell to its neighbor, and $\Delta_i$ is the D8 cellular distance. Here, the local flow depth acts as a resistance term -- as $\theta$ gets larger, the local routing weights have an increasingly large dependence on the water depth of the neighboring cells. By altering the two parameters $\gamma$ and $\theta$, the user has control over the randomness with which the particles are routed. For a full description of this methodology, see @Liang2015a; @Liang2015b and the *dorado* documentation.
+wherein $F$ are the routing directions, $d_i$ represents the unit vector from the current cell to its neighbor, and $\Delta_i$ is the D8 cellular distance. Here, the local flow depth acts as a resistance term -- as $\theta$ gets larger, the local routing weights have an increasingly large dependence on the water depth of the neighboring cells. By altering the two parameters $\gamma$ and $\theta$, the user has control over the randomness with which the particles are routed. For a full description of this methodology, see Liang et al. [-@Liang2015a; -@Liang2015b] and the *dorado* documentation.
 
[…]
