### 2020-10-22 #2585 @kbarnhart
@elbeejay and co-authors, congratulations on making it to the end of the JOSS review process. I think the software you submitted was already in great shape, but through the thoughtful comments of @dbuscombe-usgs and @gassmoeller it has been made even better. I want to thank all authors and reviewers for your contribution to this review process.

At this point @elbeejay could you:
- [ ] Make a tagged release of your software, and report the version tag of the archived version here.
