### 2020-10-22 #2585 @elbeejay
Thanks for all of your help through this process @kbarnhart. I think I've managed to do the items listed above.

- Archived version tag: v2.1.0-JOSS
- Zenodo archive: [https://zenodo.org/record/4118342](https://zenodo.org/record/4118342)
- Zenodo DOI: 10.5281/zenodo.4118342
