### 2020-10-22 #2585 @kbarnhart
@elbeejay this is now in the hands of the JOSS handling Editor in Chief who manages final publication.
